// app/src/api/room/RoomClient.ts
export interface BoardView {
  seats: string[];
  turn_seat: number;
  pending_dice: number;
  finished: boolean;
  winner: string;
  tokens: TokenView[];
}
// ...
export interface TokenView {
  seat: number;
  token_index: number;
  steps: number;
}
// ...
export class RoomClient {
  private _mero: MeroJs;
  private _contextId: string;
  private _executorPublicKey: string;

  constructor(mero: MeroJs, contextId: string, executorPublicKey: string) {
    this._mero = mero;
    this._contextId = contextId;
    this._executorPublicKey = executorPublicKey;
  }

  /**
   * init
   */
  public async init(params: { directory_context_id: string; room_id: string; min_members_to_start: number }): Promise<void> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'init', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as void;
  }

  /**
   * join_room
   */
  public async joinRoom(): Promise<number> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'join_room', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as number;
  }

  /**
   * member_count
   */
  public async memberCount(): Promise<number> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'member_count', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as number;
  }

  /**
   * min_members
   */
  public async minMembers(): Promise<number> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'min_members', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as number;
  }

  /**
   * is_started
   */
  public async isStarted(): Promise<boolean> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'is_started', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as boolean;
  }

  /**
   * roll_dice
   */
  public async rollDice(): Promise<number> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'roll_dice', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as number;
  }

  /**
   * move_token
   */
  public async moveToken(params: { token_index: number }): Promise<void> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'move_token', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as void;
  }

  /**
   * get_board
   */
  public async getBoard(): Promise<BoardView> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'get_board', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as BoardView;
  }

}